Fix a pointer truncation on x64 builds.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7057 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/WndNeighbours.cpp
@@ -362,7 +362,7 @@ CNeighbour* CNeighboursWnd::GetItem(int nItem)
 {
 	if ( m_wndList.GetItemState( nItem, LVIS_SELECTED ) )
 	{
-		return Neighbours.Get( static_cast< DWORD >( m_wndList.GetItemData( nItem ) ) );
+		return Neighbours.Get( m_wndList.GetItemData( nItem ) );
 	}
 	
 	return NULL;
